Hack application.yml to get test working other than workspace deletion, PF-612
