+++ workspace-manager-clienttests/src/main/resources/configs/integration/PrivateControlledAiNotebookInstanceLifecycle.json
@@ -1,0 +1,18 @@
+{
+  "name": "PrivateControlledAiNotebookInstanceLifecycle",
+  "description": "CRUD user-private GCS bucket tests.",
+  "serverSpecificationFile": "workspace-dev.json",
+  "kubernetes": {},
+  "application": {},
+  "testScripts": [
+    {
+      "name": "PrivateControlledAiNotebookInstanceLifecycle",
+      "numberOfUserJourneyThreadsToRun": 1,
+      "userJourneyThreadPoolSize": 2,
+      "expectedTimeForEach": 16,
+      "expectedTimeForEachUnit": "MINUTES",
+      "parameters": ["wm-default-spend-profile"]
+    }
+  ],
+  "testUserFiles": ["william.json", "bella.json", "elijah.json"]
+}
